refactor: LotteryAnnouncer -> Lotteries 로 이관
* LotteryAnnouncer의 로직을 Lotteries로 이관
* 필요한 data형식을 data class로 선언 -> Profit
* 호출부 수정
* 테스트 이관 LotteryAnnouncerTest -> LotteriesTest
