+++ src/main/kotlin/lotto/domain/Profit.kt
@@ -1,0 +1,6 @@
+package lotto.domain
+
+data class Profit(
+    val priceGroupedLotteries: Map<Price, Lotteries>,
+    val earningRate: Float
+)
